Date Scraping, PEP8 naming
Date scraping is now functional, and GroupBuy Objects now contain a meaningful attribute relating to the end of the group buy, whether it is date docused or unit limited. Not perfect, as it won't catch "until production capacity reached" or "until sold through," but it works to get the dates for GBs.

Most, if not all, variable and method names have been refactored to adhere to PEP8 guidlines.

Next in the pipeline are useful docstrings,  image links, and purchase types.